Make the server work remotely on other computers
Previously, the server would seem to only work on the local machine with
the localhost address. Now I can connect to the markdown page with that,
or the ip address of the computer.

This allows me to work on markdown over ssh, and be able to view the
file in my local browser over an internet connection (setting up some
port forwarding allows us to also do this outside of the same network).

I also made it so that the server did not automatically launch a web
browser. I don't want it launching a web browser over ssh so I just
removed it completely. It could cause problems if the machine for
instance does not have a browser to open it with, and the socket only
allows for one connection, so if my server is connected to the socket,
my local web browser cannot connect.

I would like to (in the future) be able to tell vim that I want to
automatically open a browser, and then that will tell the server to do
this but this works for my needs right now.